Unit test refactoring
