+++ jobs-api-server/src/main/java/org/zowe/jobs/services/zosmf/AbstractZosmfJobsRequestRunner.java
@@ -7,7 +7,7 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services;
+package org.zowe.jobs.services.zosmf;
 
 import com.google.gson.JsonObject;
