+++ jobs-api-server/src/main/java/org/zowe/jobs/services/zosmf/GetJobFilesZosmfRequestRunner.java
@@ -7,8 +7,9 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services;
+package org.zowe.jobs.services.zosmf;
 
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
 import lombok.extern.slf4j.Slf4j;
@@ -19,27 +20,32 @@ import org.apache.http.client.methods.RequestBuilder;
 import org.zowe.api.common.connectors.zosmf.ZosmfConnector;
 import org.zowe.api.common.exceptions.ZoweApiRestException;
 import org.zowe.api.common.utils.ResponseUtils;
-import org.zowe.jobs.exceptions.JobFileIdNotFoundException;
 import org.zowe.jobs.exceptions.JobIdNotFoundException;
 import org.zowe.jobs.exceptions.JobNameNotFoundException;
 import org.zowe.jobs.model.JobFile;
-import org.zowe.jobs.model.JobFileContent;
 
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
+import java.util.List;
 
 @Slf4j
-public class GetJobFileContentZosmfRequestRunner extends AbstractZosmfRequestRunner<JobFileContent> {
+public class GetJobFilesZosmfRequestRunner extends AbstractZosmfRequestRunner<List<JobFile>> {
 
     private String jobName;
     private String jobId;
-    private String fileId;
 
-    public GetJobFileContentZosmfRequestRunner(String jobName, String jobId, String fileId) {
+    public GetJobFilesZosmfRequestRunner(String jobName, String jobId) {
         this.jobName = jobName;
         this.jobId = jobId;
-        this.fileId = fileId;
+    }
+
+    @Override
+    RequestBuilder prepareQuery(ZosmfConnector zosmfConnector) throws URISyntaxException {
+        String urlPath = String.format("restjobs/jobs/%s/%s/files", jobName, jobId); //$NON-NLS-1$
+        URI requestUrl = zosmfConnector.getFullUrl(urlPath);
+        return RequestBuilder.get(requestUrl);
     }
 
     @Override
@@ -48,18 +54,15 @@ public class GetJobFileContentZosmfRequestRunner extends AbstractZosmfRequestRun
     }
 
     @Override
-    RequestBuilder prepareQuery(ZosmfConnector zosmfconnector) throws URISyntaxException {
-        String urlPath = String.format("restjobs/jobs/%s/%s/files/%s/records", jobName, jobId, fileId); //$NON-NLS-1$
-        URI requestUrl = zosmfconnector.getFullUrl(urlPath);
-        return RequestBuilder.get(requestUrl);
+    List<JobFile> getResult(HttpResponse response) throws IOException {
+        JsonElement jsonResponse = ResponseUtils.getEntityAsJson(response);
+        List<JobFile> jobFiles = new ArrayList<>();
+        for (JsonElement jsonElement : jsonResponse.getAsJsonArray()) {
+            jobFiles.add(getJobFileFromJson(jsonElement.getAsJsonObject()));
+        }
+        return jobFiles;
     }
 
-    @Override
-    JobFileContent getResult(HttpResponse response) throws IOException {
-        return new JobFileContent(ResponseUtils.getEntity(response));
-    }
-
-    // TODO NOW - review the createExceptions to look for common behaviour
     @Override
     ZoweApiRestException createException(JsonObject jsonResponse, int statusCode) {
         if (statusCode == HttpStatus.SC_BAD_REQUEST) {
@@ -67,9 +70,6 @@ public class GetJobFileContentZosmfRequestRunner extends AbstractZosmfRequestRun
                 String zosmfMessage = jsonResponse.get("message").getAsString();
                 if (String.format("No job found for reference: '%s(%s)'", jobName, jobId).equals(zosmfMessage)) {
                     return new JobNameNotFoundException(jobName, jobId);
-                } else if (String.format("Job '%s(%s)' does not contain spool file id %s", jobName, jobId, fileId)
-                    .equals(zosmfMessage)) {
-                    return new JobFileIdNotFoundException(jobName, jobId, fileId);
                 }
             }
         } else if (statusCode == HttpStatus.SC_INTERNAL_SERVER_ERROR) {
