+++ jobs-api-server/src/main/java/org/zowe/jobs/services/zosmf/PurgeJobZosmfRequestRunner.java
@@ -7,9 +7,8 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services;
+package org.zowe.jobs.services.zosmf;
 
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
 import lombok.extern.slf4j.Slf4j;
@@ -19,42 +18,38 @@ import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.RequestBuilder;
 import org.zowe.api.common.connectors.zosmf.ZosmfConnector;
 import org.zowe.api.common.exceptions.ZoweApiRestException;
-import org.zowe.api.common.utils.ResponseUtils;
-import org.zowe.jobs.exceptions.JobIdNotFoundException;
 import org.zowe.jobs.exceptions.JobNameNotFoundException;
-import org.zowe.jobs.model.Job;
 
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 
 @Slf4j
-public class GetJobZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<Job> {
+public class PurgeJobZosmfRequestRunner extends AbstractZosmfRequestRunner<Void> {
 
     private String jobName;
     private String jobId;
 
-    public GetJobZosmfRequestRunner(String jobName, String jobId) {
+    public PurgeJobZosmfRequestRunner(String jobName, String jobId) {
         this.jobName = jobName;
         this.jobId = jobId;
     }
 
     @Override
-    RequestBuilder prepareQuery(ZosmfConnector zosmfconnector) throws URISyntaxException {
-        String urlPath = String.format("restjobs/jobs/%s/%s", jobName, jobId); //$NON-NLS-1$
-        URI requestUrl = zosmfconnector.getFullUrl(urlPath);
-        return RequestBuilder.get(requestUrl);
-    }
-
-    @Override
     int[] getSuccessStatus() {
-        return new int[] { HttpStatus.SC_OK };
+        return new int[] { HttpStatus.SC_ACCEPTED };
     }
 
     @Override
-    Job getResult(HttpResponse response) throws IOException {
-        JsonElement jsonResponse = ResponseUtils.getEntityAsJson(response);
-        return getJobFromJson(jsonResponse.getAsJsonObject());
+    RequestBuilder prepareQuery(ZosmfConnector zosmfConnector) throws URISyntaxException {
+        String urlPath = String.format("restjobs/jobs/%s/%s", jobName, jobId); //$NON-NLS-1$
+        URI requestUrl = zosmfConnector.getFullUrl(urlPath);
+        return RequestBuilder.delete(requestUrl);
+    }
+
+    @Override
+    Void getResult(HttpResponse response) throws IOException {
+        return null;
     }
 
     @Override
@@ -63,14 +58,7 @@ public class GetJobZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<Job
             if (jsonResponse.has("message")) {
                 String zosmfMessage = jsonResponse.get("message").getAsString();
                 if (String.format("No job found for reference: '%s(%s)'", jobName, jobId).equals(zosmfMessage)) {
-                    throw new JobNameNotFoundException(jobName, jobId);
-                }
-            }
-        } else if (statusCode == HttpStatus.SC_INTERNAL_SERVER_ERROR) {
-            if (jsonResponse.has("message")) {
-                String zosmfMessage = jsonResponse.get("message").getAsString();
-                if (String.format("Failed to lookup job %s(%s)", jobName, jobId).equals(zosmfMessage)) {
-                    throw new JobIdNotFoundException(jobName, jobId);
+                    return new JobNameNotFoundException(jobName, jobId);
                 }
             }
         }
