+++ jobs-api-server/src/test/java/org/zowe/jobs/services/zosmf/PurgeJobZosmfRequestRunnerTest.java
@@ -1,0 +1,47 @@
+package org.zowe.jobs.services.zosmf;
+
+import org.apache.http.HttpResponse;
+import org.apache.http.HttpStatus;
+import org.apache.http.client.methods.RequestBuilder;
+import org.junit.Test;
+import org.zowe.jobs.exceptions.JobNameNotFoundException;
+
+import static org.mockito.Mockito.when;
+
+public class PurgeJobZosmfRequestRunnerTest extends AbstractZosmfJobsRequestRunnerTest {
+
+    @Test
+    public void purge_job_string_should_call_zosmf_correctly() throws Exception {
+        String jobName = "AJOB";
+        String jobId = "Job12345";
+
+        HttpResponse response = mockResponse(HttpStatus.SC_ACCEPTED);
+
+        RequestBuilder requestBuilder = mockDeleteBuilder(String.format("restjobs/jobs/%s/%s", jobName, jobId));
+
+        when(zosmfConnector.request(requestBuilder)).thenReturn(response);
+
+        new PurgeJobZosmfRequestRunner(jobName, jobId).run(zosmfConnector);
+
+        verifyInteractions(requestBuilder);
+    }
+
+    @Test
+    public void purge_job_for_non_existing_job_should_throw_exception() throws Exception {
+        String jobName = "ATLJ5000";
+        String jobId = "JOB21489";
+
+        Exception expectedException = new JobNameNotFoundException(jobName, jobId);
+
+        HttpResponse response = mockJsonResponse(HttpStatus.SC_BAD_REQUEST,
+                loadTestFile("zosmf_getJob_noJobNameResponse.json"));
+
+        RequestBuilder requestBuilder = mockDeleteBuilder(String.format("restjobs/jobs/%s/%s", jobName, jobId));
+
+        when(zosmfConnector.request(requestBuilder)).thenReturn(response);
+
+        shouldThrow(expectedException, () -> new PurgeJobZosmfRequestRunner(jobName, jobId).run(zosmfConnector));
+        verifyInteractions(requestBuilder);
+    }
+
+}
